Added getLogger
